+++ api/client/video/index.js
@@ -1,6 +1,7 @@
 let router = module.exports = require('express').Router();
 
 router.get('/', require('./get.get-video-list'));
+// router.get('/recommend', require('./get.fetch-recommend-video-list'));
 router.get('/:videoId', require('./get.get-video-profile'));
 router.get('/channel', require('./get.get-channel-list'));
 router.get('/channel/:channelId', require('./get.get-channel-profile'));