+++ utils/transform-query.js
@@ -14,8 +14,16 @@ module.exports = (query, data, strict = false) => {
           else {
             delete query[key];
           }
-
           break;
+
+        case 'StringArray':
+          if (typeof value === 'string') {
+            query[key] = value.split(',').map(s => s.trim());
+          }
+
+          else {
+            delete query[key];
+          }
 
         default:
           //  ...